Organise slider components
