+++ components/Fields/Slider/Rail.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const railStyle = {
+  position: 'absolute',
+  width: '100%',
+  height: 10,
+  marginTop: 35,
+  borderRadius: 5,
+  backgroundColor: '#8B9CB6',
+};
+const Rail = () => (
+  <div style={railStyle} />
+);
+
+export default Rail;
