+++ components/Fields/Slider/Tick.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Tick = ({ value }) => (
+  <div>{value}</div>
+);
+
+export default Tick;
